update root dispatch context
