+++ python/tvm/autotvm/__init__.py
@@ -24,6 +24,6 @@ from . import tophub
 # some shortcuts
 from .measure import measure_option, MeasureInput, MeasureResult, MeasureErrorNo, use_rpc
 from .tuner import callback
-from .task import template, get_config, create, ConfigSpace, ConfigEntity
-from .record import ApplyHistoryBest as apply_history_best
+from .task import template, get_config, create, ConfigSpace, ConfigEntity, \
+    ApplyHistoryBest as apply_history_best
 from .env import GLOBAL_SCOPE
