+++ python/tvm/autotvm/tophub.py
@@ -9,26 +9,16 @@ import logging
 import os
 import json
 
-from .record import ApplyHistoryBest
+from .task import ROOT_DISPATCH_CONTEXT
 from ..contrib.util import tempdir
 from ..contrib.download import download
 
 AUTOTVM_TOPHUB_ROOT_PATH = os.path.join(os.path.expanduser('~'), ".tvm", "tophub")
 
 
-class EmptyContext(object):
-    """An empty context"""
-    def __enter__(self):
-        pass
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        pass
-
-
-def context(target, extra_files=None):
-    """Return the dispatch context with pre-tuned parameters.
-    The corresponding downloaded *.log files under tophub root path will be loaded.
-    Users can also add their own files in argument `extra_files`.
+def load(target, extra_files=None):
+    """Load pre-tuned parameters to the root dispatch context.
+    The corresponding downloaded *.log files under TopHub root path will be loaded.
 
     Parameters
     ----------
@@ -36,25 +26,28 @@ def context(target, extra_files=None):
         The compilation target
     extra_files: list of str, optional
         Extra log files to load
+
+    Note
+    ----
+    If the current dispatch context is not root dispatch context,
+    this function won't affect the current dispatch context.
     """
     rootpath = AUTOTVM_TOPHUB_ROOT_PATH
-    best_context = ApplyHistoryBest([])
+    root_context = ROOT_DISPATCH_CONTEXT
 
     big_target = str(target).split()[0]
     if os.path.isfile(os.path.join(rootpath, big_target + ".log")):
-        best_context.load(os.path.join(rootpath, big_target + ".log"))
+        root_context.load(os.path.join(rootpath, big_target + ".log"))
 
     for opt in target.options:
         if opt.startswith("-device"):
             model = opt[8:]
             if os.path.isfile(os.path.join(rootpath, model) + ".log"):
-                best_context.load(os.path.join(rootpath, model) + ".log")
+                root_context.load(os.path.join(rootpath, model) + ".log")
 
     if extra_files:
         for filename in extra_files:
-            best_context.load(filename)
-
-    return best_context
+            root_context.load(filename)
 
 
 def download_package(backend):
